Fleshing out the reservation menu
Got rid of code in Utilities.cpp that I copypasted  from the old project.

Fleshed out the reservation menu, replaced if/else if/else chains from the old project with switch cases.

Note: main.cpp is  turning into a mess of nested switch cases. Maybe there is a way to clean it up?