Working on the Special Requests Menu
Figuring out how to get this to work. I figure the best way to approach this is to save any special requests to a file in the Reservation Menu part of the top menu, and then just read those out in the Advanced Options Menu.

So, I'm using an std::vector to hold the requests, and then I save those to a file.

Maybe there is a way to just read the file line-by-line in the Advanced Options Menu? It doesn't make any sense to me to load another vector in that part of the program, and since these are vectors of strings, this is an expensive operation.